refactor alloc. remove global (tls) SYCL context and device.
