+++ tutorials/common/CMakeLists.txt
@@ -5,6 +5,7 @@ IF (EMBREE_TUTORIALS_GLFW)
   ADD_SUBDIRECTORY(imgui)
 ENDIF()
 
+ADD_SUBDIRECTORY(alloc)
 ADD_SUBDIRECTORY(tutorial)
 ADD_SUBDIRECTORY(scenegraph)
 ADD_SUBDIRECTORY(lights)
